+++ crates/cli-support/src/descriptors.rs
@@ -47,6 +47,11 @@ impl WasmBindgenDescriptorsSection {
         interpreter: &mut Interpreter,
     ) -> Result<(), Error> {
         let mut to_remove = Vec::new();
+
+        if let Some(id) = interpreter.skip_interpret() {
+            to_remove.push(id);
+        }
+
         for export in module.exports.iter() {
             let prefix = "__wbindgen_describe_";
             if !export.name.starts_with(prefix) {